# Conflicts:
#	sik.sql
#	src/bridging/InhealthTagihan.form
#	src/bridging/InhealthTagihan.java
#	src/fungsi/akses.java
#	src/grafikanalisa/GrafikPenerimaanObatPerbulan.form
#	src/grafikanalisa/GrafikPenerimaanObatPerbulan.java
#	src/inventory/DlgCariObat2.form
#	src/inventory/DlgCariPenjualan.java
#	src/inventory/DlgCariPiutang.java
#	src/inventory/DlgDaftarPermintaanResep.java
#	src/inventory/DlgInputStok.java
#	src/inventory/DlgMutasiBarang.java
#	src/inventory/DlgPengambilanUTD.java
#	src/inventory/DlgPengeluaranApotek.java
#	src/inventory/DlgPenggunaObat.java
#	src/inventory/DlgPenjualan.java
#	src/inventory/DlgPiutang.java
#	src/inventory/DlgProyeksiJual.java
#	src/inventory/DlgResepPulang.java
#	src/inventory/DlgReturBeli.java
#	src/inventory/DlgReturJual.java
#	src/inventory/DlgReturPiutang.java
#	src/inventory/DlgSatuan.form
#	src/inventory/DlgSatuan.java
#	src/inventory/HibahObatBHP.java
#	src/keuangan/DlgPemasukanLain.java
#	src/keuangan/DlgPengeluaranHarian.java
#	src/permintaan/DlgCariPermintaanRadiologi.java
#	src/permintaan/DlgPermintaanLaboratorium.java
#	src/permintaan/DlgPermintaanRadiologi.java
#	src/rekammedis/DataTriaseIGD.java
#	src/rekammedis/RMDataMonitoringAsuhanGizi.form
#	src/rekammedis/RMDataMonitoringAsuhanGizi.java
#	src/setting/DlgUpdateUser.java
#	src/setting/DlgUser.java
#	src/simrskhanza/DlgCariPeriksaRadiologi.java
#	src/simrskhanza/DlgIGD.form
#	src/simrskhanza/DlgIGD.java
#	src/simrskhanza/DlgKamarInap.java
#	src/simrskhanza/DlgKasirRalan.form
#	src/simrskhanza/DlgKasirRalan.java
#	src/simrskhanza/DlgPasien.java
#	src/simrskhanza/DlgRawatInap.java
#	src/simrskhanza/DlgReg.form
#	src/simrskhanza/DlgReg.java
#	src/simrskhanza/frmUtama.form
#	src/simrskhanza/frmUtama.java

+++ src/rekammedis/RMDataMonitoringAsuhanGizi.java
@@ -102,9 +102,9 @@ public final class RMDataMonitoringAsuhanGizi extends javax.swing.JDialog {
 
         TNoRw.setDocument(new batasInput((byte)17).getKata(TNoRw));
         nip.setDocument(new batasInput((byte)20).getKata(nip));
-        Monitoring.setDocument(new batasInput(60).getKata(Monitoring));
-        Evaluasi.setDocument(new batasInput(200).getKata(Evaluasi));
-        TCari.setDocument(new batasInput(100).getKata(TCari));
+        Monitoring.setDocument(new batasInput((int)60).getKata(Monitoring));
+        Evaluasi.setDocument(new batasInput((int)200).getKata(Evaluasi));
+        TCari.setDocument(new batasInput((int)100).getKata(TCari));
         
         if(koneksiDB.CARICEPAT().equals("aktif")){
             TCari.getDocument().addDocumentListener(new javax.swing.event.DocumentListener(){
